feat: types폴더 내 타입파일명 XXXType으로 통일
