+++ src/pages/auth-pages/LoginPage.tsx
@@ -14,13 +14,12 @@ import {
   PAGE_FIND_PW,
   PAGE_SIGNUP_SELECT,
 } from 'constants/auth/path'
-import { Message, SigninInputValue } from 'types/auth'
+import { Message, ResultWithData, SigninInputValue } from 'types/authType'
 import { authService } from 'services/auth/authService'
 import MessageModal from 'components/common/MessageModal'
 import { loginUser } from 'redux/store/userSlice'
 import { useDispatch } from 'react-redux'
 import { userService } from 'services/auth/userService'
-import { ResultWithData } from 'types/setting'
 
 const LoginPage = () => {
   const [inputValue, setInputValue] = useState<SigninInputValue>({
@@ -74,7 +73,6 @@ const LoginPage = () => {
 
   const submitSigninForm = async (e: MouseEvent<HTMLFormElement>) => {
     e.preventDefault()
-
     const { result, message } = await authService.signin(email, password)
     const path: string = result === 'success' ? '/' : ''
     // TODO: 아파트 인증 여부에 따라, 메인 홈(미인증 사용자) 또는 커뮤니티 홈(인증 사용자)으로 리다이렉트 추가 예정
