+++ src/pages/auth-pages/SignupLocalPage.tsx
@@ -4,7 +4,7 @@ import { Button } from 'styles/reusable-style/elementStyle'
 import AuthLayout from 'components/auth/AuthLayout'
 import SignupInputArea from 'components/auth/SignupInputArea'
 import TermsAndConditionArea from 'components/auth/TermsAndConditionArea'
-import { Message, SignupInputValue, TermsAndConditionsValue } from 'types/auth'
+import { Message, SignupInputValue, TermsAndConditionsValue } from 'types/authType'
 import MessageModal from 'components/common/MessageModal'
 import { authService } from 'services/auth/authService'
 import { PAGE_LOGIN } from 'constants/auth/path'
