+++ src/pages/setting-pages/ChangeImagePage.tsx
@@ -2,7 +2,7 @@ import { useState, useRef, ChangeEvent, MouseEvent } from 'react'
 import { useNavigate } from 'react-router-dom'
 import { styled } from 'styled-components'
 import AuthLayout from 'components/auth/AuthLayout'
-import { Message } from 'types/auth'
+import { Message } from 'types/authType'
 import { Button, Img } from 'styles/reusable-style/elementStyle'
 import { PAGE_SETTING } from 'constants/setting/path'
 import MessageModal from 'components/common/MessageModal'
