+++ src/pages/setting-pages/ChangePwPage.tsx
@@ -3,12 +3,12 @@ import AuthLayout from 'components/auth/AuthLayout'
 import SignupInput from 'components/auth/SignupInput'
 import { styled } from 'styled-components'
 import { Button } from 'styles/reusable-style/elementStyle'
-import { ChangePwInputValue } from 'types/setting'
+import { ChangePwInputValue } from 'types/settingType'
 import { useNavigate } from 'react-router-dom'
 import { PAGE_SETTING } from 'constants/setting/path'
 import { userService } from 'services/auth/userService'
 import { signupValidation } from 'constants/auth/signupValidation'
-import { Message } from 'types/auth'
+import { Message } from 'types/authType'
 import MessageModal from 'components/common/MessageModal'
 
 const ChangePwPage = () => {
