+++ src/pages/setting-pages/MyArticlePage.tsx
@@ -3,7 +3,7 @@ import { userService } from 'services/auth/userService'
 import { styled } from 'styled-components'
 import { ShadowBox } from 'styles/reusable-style/elementStyle'
 import { Container, Inner } from 'styles/reusable-style/layoutStyle'
-import { MyArticle, ResultWithData } from 'types/setting'
+import { MyArticle, ResultWithData } from 'types/settingType'
 import { timeAgo } from 'utils/timeAgo'
 import { SIZE_OPTION } from 'constants/setting/pagination'
 import Pagination from 'components/common/Pagination'
