+++ src/types/settingType.ts
@@ -48,18 +48,3 @@ export type ResultWithData = {
   result: 'success'
   data: any
 }
-
-export type Result = {
-  result: 'success' | 'fail'
-  message: string
-}
-
-export type ContactInputValue = {
-  dataCollectAgree: boolean
-  name: string
-  nonAuth: boolean
-  email: string
-  code: string
-  title: string
-  content: string
-}
